+++ autoload/omnipytent/__init__.py
@@ -3,6 +3,8 @@ __all__ = ['task', 'CMD', 'FN', 'VAR', 'BANG', 'SH']
 from .dsl import task
 from .execution import CMD, FN, VAR, BANG, SH
 
+from . import _extending as __extending
+
 try:
     from .execution import TERMINAL_TAB, TERMINAL_PANEL
     __all__.extend(['TERMINAL_TAB', 'TERMINAL_PANEL'])