+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Windows support.
 - `TERMINAL_PANEL` and `TERMINAL_TAB` support for Vim 8.
+- `omnipytent.ext` for loading module from `omnipytent` directories in Vim plugins.
 
 ## 0.1.0 - 2016-08-26
 ### Added